+++ src/main/java/de/jost_net/JVerein/gui/control/AbrechnungslaufControl.java
@@ -18,34 +18,27 @@ package de.jost_net.JVerein.gui.control;
 
 import java.rmi.RemoteException;
 import java.util.Date;
-import java.util.Map.Entry;
 import de.jost_net.JVerein.Einstellungen;
 import de.jost_net.JVerein.Einstellungen.Property;
 import de.jost_net.JVerein.gui.action.BuchungAction;
 import de.jost_net.JVerein.gui.action.EditAction;
 import de.jost_net.JVerein.gui.dialogs.TabelleSpaltenAuswahlDialog;
-import de.jost_net.JVerein.gui.dialogs.AbstractPartExportDialog.ExportArt;
-import de.jost_net.JVerein.gui.formatter.AbrechnungsmodusFormatter;
 import de.jost_net.JVerein.gui.formatter.BuchungsartFormatter;
 import de.jost_net.JVerein.gui.formatter.BuchungsklasseFormatter;
 import de.jost_net.JVerein.gui.formatter.IBANFormatter;
 import de.jost_net.JVerein.gui.formatter.JaNeinFormatter;
 import de.jost_net.JVerein.gui.formatter.SollbuchungFormatter;
-import de.jost_net.JVerein.gui.menu.AbrechnungslaufMenu;
 import de.jost_net.JVerein.gui.menu.BuchungAbrechnugslaufMenu;
 import de.jost_net.JVerein.gui.menu.LastschriftMenu;
 import de.jost_net.JVerein.gui.menu.SollbuchungMenu;
 import de.jost_net.JVerein.gui.menu.ZusatzbetraegeMenu;
-import de.jost_net.JVerein.gui.parts.AutoUpdateTablePart;
 import de.jost_net.JVerein.gui.parts.BetragSummaryTablePart;
 import de.jost_net.JVerein.gui.parts.BuchungListTablePart;
 import de.jost_net.JVerein.gui.parts.JVereinTablePart;
-import de.jost_net.JVerein.gui.view.AbrechnungslaufDetailView;
 import de.jost_net.JVerein.gui.view.LastschriftDetailView;
 import de.jost_net.JVerein.gui.view.SollbuchungDetailView;
 import de.jost_net.JVerein.gui.view.ZusatzbetragDetailView;
 import de.jost_net.JVerein.keys.Abrechnungsmodi;
-import de.jost_net.JVerein.keys.Filter;
 import de.jost_net.JVerein.keys.SplitbuchungTyp;
 import de.jost_net.JVerein.keys.VorlageTyp;
 import de.jost_net.JVerein.keys.Zahlungsweg;
@@ -60,7 +53,6 @@ import de.jost_net.JVerein.util.JVDateFormatTTMMJJJJ;
 import de.jost_net.JVerein.util.VorlageUtil;
 import de.willuhn.datasource.pseudo.PseudoIterator;
 import de.willuhn.datasource.rmi.DBIterator;
-import de.willuhn.datasource.rmi.DBService;
 import de.willuhn.jameica.gui.AbstractView;
 import de.willuhn.jameica.gui.formatter.CurrencyFormatter;
 import de.willuhn.jameica.gui.formatter.DateFormatter;
@@ -74,12 +66,10 @@ import de.willuhn.jameica.system.OperationCanceledException;
 import de.willuhn.logging.Logger;
 import de.willuhn.util.ApplicationException;
 
-public class AbrechnungslaufControl extends FilterControl implements Savable
+public class AbrechnungslaufControl extends VorZurueckControl implements Savable
 {
 
   private Abrechnungslauf abrl;
-
-  private JVereinTablePart abrechnungslaufList;
 
   private LabelInput datum;
 
@@ -303,99 +293,21 @@ public class AbrechnungslaufControl extends FilterControl implements Savable
   public JVereinTablePart getTablePart()
       throws RemoteException, ApplicationException
   {
-    if (abrechnungslaufList != null)
+    switch (selectedTab)
     {
-      return abrechnungslaufList;
+      case TAB_BUCHUNGEN:
+        return getBuchungList();
+      case TAB_LASTSCHRIFTEN:
+        return getLastschriftList();
+      case TAB_SOLLBUCHUNGEN:
+        return getSollbuchungList();
+      case TAB_ZUSATZBETRAEGE:
+        return getZusatzbetraegeList();
+      case TAB_ALLEBUCHUNGEN:
+        return getAlleBuchungList();
+      default:
+        return null;
     }
-
-    abrechnungslaufList = new AutoUpdateTablePart(getAbrechnungslaeufe(),
-        new EditAction(AbrechnungslaufDetailView.class));
-    abrechnungslaufList.addColumn("Nr", "nr");
-    if ((Boolean) Einstellungen.getEinstellung(Property.ABRLABSCHLIESSEN))
-    {
-      abrechnungslaufList.addColumn("Abgeschlossen", "abgeschlossen",
-          o -> (Boolean) o ? "\uD83D\uDD12" : "");
-    }
-    abrechnungslaufList.addColumn("Datum", "datum",
-        new DateFormatter(new JVDateFormatTTMMJJJJ()));
-    abrechnungslaufList.addColumn("Modus", "modus",
-        new AbrechnungsmodusFormatter(), false, Column.ALIGN_LEFT);
-    abrechnungslaufList.addColumn("Fälligkeit", "faelligkeit",
-        new DateFormatter(new JVDateFormatTTMMJJJJ()));
-    abrechnungslaufList.addColumn("Stichtag", "stichtag",
-        new DateFormatter(new JVDateFormatTTMMJJJJ()));
-    abrechnungslaufList.addColumn("Eintrittsdatum", "eingabedatum",
-        new DateFormatter(new JVDateFormatTTMMJJJJ()));
-    abrechnungslaufList.addColumn("Austrittsdatum", "austrittsdatum",
-        new DateFormatter(new JVDateFormatTTMMJJJJ()));
-    abrechnungslaufList.addColumn("Zahlungsgrund", "zahlungsgrund");
-    if ((Boolean) Einstellungen.getEinstellung(Property.ZUSATZBETRAG))
-    {
-      abrechnungslaufList.addColumn("Zusatzbeträge", "zusatzbetraege",
-          new JaNeinFormatter());
-    }
-    if ((Boolean) Einstellungen.getEinstellung(Property.KURSTEILNEHMER))
-    {
-      abrechnungslaufList.addColumn("Kursteilnehmer", "kursteilnehmer",
-          new JaNeinFormatter());
-    }
-    abrechnungslaufList
-        .setContextMenu(new AbrechnungslaufMenu(abrechnungslaufList));
-    abrechnungslaufList.setAction(
-        new EditAction(AbrechnungslaufDetailView.class, abrechnungslaufList));
-    VorZurueckControl.setObjektListe(null, null);
-
-    return abrechnungslaufList;
-  }
-
-  @Override
-  protected void TabRefresh() throws ApplicationException
-  {
-    if (abrechnungslaufList == null)
-    {
-      return;
-    }
-    try
-    {
-      DBIterator<Abrechnungslauf> abrechnungslaeufe = getAbrechnungslaeufe();
-      abrechnungslaufList.removeAll();
-      while (abrechnungslaeufe.hasNext())
-      {
-        abrechnungslaufList.addItem(abrechnungslaeufe.next());
-      }
-      abrechnungslaufList.sort();
-    }
-    catch (RemoteException e1)
-    {
-      Logger.error("Fehler", e1);
-    }
-  }
-
-  private DBIterator<Abrechnungslauf> getAbrechnungslaeufe()
-      throws RemoteException, ApplicationException
-  {
-    DBService service = Einstellungen.getDBService();
-    DBIterator<Abrechnungslauf> abrechnungslaeufe = service
-        .createList(Abrechnungslauf.class);
-
-    for (Entry<Filter, Object> entry : getFilter().entrySet())
-    {
-      Object value = entry.getValue();
-      switch (entry.getKey())
-      {
-        case DATUM_VON:
-          abrechnungslaeufe.addFilter("datum >= ?", value);
-          break;
-        case DATUM_BIS:
-          abrechnungslaeufe.addFilter("datum <= ?", value);
-          break;
-        default:
-          throw new ApplicationException(
-              "Filter nicht implementiert: " + entry.getKey().getAnzeigeText());
-      }
-    }
-    abrechnungslaeufe.setOrder("ORDER BY datum DESC");
-    return abrechnungslaeufe;
   }
 
   @SuppressWarnings("unchecked")
@@ -673,126 +585,93 @@ public class AbrechnungslaufControl extends FilterControl implements Savable
     }, "Spalten auswählen");
   }
 
-  public PanelButton exportDetailButton(ExportArt art)
-      throws ApplicationException
-  {
-
-    return new PanelButton(
-        art.equals(ExportArt.PDF) ? "file-pdf.png" : "xsd.png", context -> {
-          final JVereinTablePart liste;
-          switch (selectedTab)
-          {
-            case TAB_BUCHUNGEN:
-              liste = buchungList;
-              break;
-            case TAB_LASTSCHRIFTEN:
-              liste = lastschriftList;
-              break;
-            case TAB_SOLLBUCHUNGEN:
-              liste = sollbuchungList;
-              break;
-            case TAB_ZUSATZBETRAEGE:
-              liste = zusatzbetraegeList;
-              break;
-            case TAB_ALLEBUCHUNGEN:
-              liste = allebuchungList;
-              break;
-            default:
-              liste = null;
-          }
-          if (liste == null)
-          {
-            throw new ApplicationException(
-                "PDF Button kann nicht erstellt werden, Tabelle ist nicht geladen.");
-          }
-
-          switch (selectedTab)
-          {
-            case TAB_BUCHUNGEN:
-              liste.export(
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_BUCHUNGEN_TITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_BUCHUNGEN_SUBTITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_BUCHUNGEN_DATEINAME,
-                      getAbrechnungslauf()),
-                  art);
-              break;
-            case TAB_LASTSCHRIFTEN:
-              liste.export(
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_LASTSCHRIFTEN2_TITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_LASTSCHRIFTEN2_SUBTITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_LASTSCHRIFTEN2_DATEINAME,
-                      getAbrechnungslauf()),
-                  art);
-              break;
-            case TAB_SOLLBUCHUNGEN:
-              liste.export(
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_SOLLBUCHUNGEN_TITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_SOLLBUCHUNGEN_SUBTITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_SOLLBUCHUNGEN_DATEINAME,
-                      getAbrechnungslauf()),
-                  art);
-              break;
-            case TAB_ZUSATZBETRAEGE:
-              liste.export(
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_ZUSATZBETRAEGE_TITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_ZUSATZBETRAEGE_SUBTITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_ZUSATZBETRAEGE_DATEINAME,
-                      getAbrechnungslauf()),
-                  art);
-              break;
-            case TAB_ALLEBUCHUNGEN:
-              liste.export(
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_ALLEBUCHUNGEN_TITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_ALLEBUCHUNGEN_SUBTITEL,
-                      getAbrechnungslauf()),
-                  VorlageUtil.getName(
-                      VorlageTyp.ABRECHNUNGSLAUF_ALLEBUCHUNGEN_DATEINAME,
-                      getAbrechnungslauf()),
-                  art);
-              break;
-          }
-        }, art.equals(ExportArt.PDF) ? "PDF" : "CSV");
-  }
-
   @Override
   protected String getTableTitle()
   {
-    return VorlageUtil.getName(VorlageTyp.ABRECHNUNGSLAEUFE_TITEL, this);
+    switch (selectedTab)
+    {
+      case TAB_BUCHUNGEN:
+        return VorlageUtil.getName(VorlageTyp.ABRECHNUNGSLAUF_BUCHUNGEN_TITEL,
+            getAbrechnungslauf());
+      case TAB_LASTSCHRIFTEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_LASTSCHRIFTEN2_TITEL,
+            getAbrechnungslauf());
+      case TAB_SOLLBUCHUNGEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_SOLLBUCHUNGEN_TITEL,
+            getAbrechnungslauf());
+      case TAB_ZUSATZBETRAEGE:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_ZUSATZBETRAEGE_TITEL,
+            getAbrechnungslauf());
+      case TAB_ALLEBUCHUNGEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_ALLEBUCHUNGEN_TITEL,
+            getAbrechnungslauf());
+      default:
+        return null;
+    }
   }
 
   @Override
   protected String getTableSubtitle()
   {
-    return VorlageUtil.getName(VorlageTyp.ABRECHNUNGSLAEUFE_SUBTITEL, this);
+    switch (selectedTab)
+    {
+      case TAB_BUCHUNGEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_BUCHUNGEN_SUBTITEL,
+            getAbrechnungslauf());
+      case TAB_LASTSCHRIFTEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_LASTSCHRIFTEN2_SUBTITEL,
+            getAbrechnungslauf());
+      case TAB_SOLLBUCHUNGEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_SOLLBUCHUNGEN_SUBTITEL,
+            getAbrechnungslauf());
+      case TAB_ZUSATZBETRAEGE:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_ZUSATZBETRAEGE_SUBTITEL,
+            getAbrechnungslauf());
+      case TAB_ALLEBUCHUNGEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_ALLEBUCHUNGEN_SUBTITEL,
+            getAbrechnungslauf());
+      default:
+        return null;
+    }
   }
 
   @Override
   protected String getTableDateiname()
   {
-    return VorlageUtil.getName(VorlageTyp.ABRECHNUNGSLAEUFE_DATEINAME, this);
+    switch (selectedTab)
+    {
+      case TAB_BUCHUNGEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_BUCHUNGEN_DATEINAME,
+            getAbrechnungslauf());
+      case TAB_LASTSCHRIFTEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_LASTSCHRIFTEN2_DATEINAME,
+            getAbrechnungslauf());
+      case TAB_SOLLBUCHUNGEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_SOLLBUCHUNGEN_DATEINAME,
+            getAbrechnungslauf());
+      case TAB_ZUSATZBETRAEGE:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_ZUSATZBETRAEGE_DATEINAME,
+            getAbrechnungslauf());
+      case TAB_ALLEBUCHUNGEN:
+        return VorlageUtil.getName(
+            VorlageTyp.ABRECHNUNGSLAUF_ALLEBUCHUNGEN_DATEINAME,
+            getAbrechnungslauf());
+      default:
+        return null;
+    }
   }
 
 }
