+++ src/main/java/de/jost_net/JVerein/gui/view/AbrechnungslaufDetailView.java
@@ -126,8 +126,9 @@ public class AbrechnungslaufDetailView extends AbstractDetailView
     buttons.addButton(new SaveButton(control));
     buttons.paint(this.getParent());
 
-    GUI.getView().addPanelButton(control.exportDetailButton(ExportArt.PDF));
-    GUI.getView().addPanelButton(control.exportDetailButton(ExportArt.CSV));
+    GUI.getView().addPanelButton(control.getProfilePanelButton());
+    GUI.getView().addPanelButton(control.exportButton(ExportArt.PDF));
+    GUI.getView().addPanelButton(control.exportButton(ExportArt.CSV));
     GUI.getView().addPanelButton(control.getDetailPanelButton());
   }
